Add SwiftLint script and image handling functions to CreateSuggestionView
- Add SwiftLint script build phases to project.pbxproj
- Set ENABLE_USER_SCRIPT_SANDBOXING to NO in project.pbxproj for multiple targets
- Create CreateSuggestionView+Functions.swift with image handling and compression functions
- Move image handling and compression logic from CreateSuggestionView.swift to CreateSuggestionView+Functions.swift
- Update CreateSuggestionView.swift to use new image handling functions and adjust view model initialization
